[buildsys: c(xx)flags -flto]
Move the link time optimisation (lto) c(xx)flags for target build from
the Makefile to the project configuration.

With recent versions of gcc incremental linking changed such that it
doesn't work anymore with lto for bare-metal systems. To avoid disabling
lto for targets entirely it has been moved to the project configuration.

Also sync cflags and cxxflags for the existing project configurations.