[buildsys: CONFIG_CODE_COVERAGE]
This change adds the CONFIG_CODE_COVERAGE flag to control if the build
shall support code coverage acquisition.

Besides controlling the respective compiler flags it also ensures the
linking of the coverage library as well as the content of the coverage
make target.
To allow the execution of tests without code coverage, the test make
target has been extracted from coverage.make.