[interrupts: foretell]
With this change the interrupt foretell mechanism is added, which allows
to register software interrupts that shall be handled upon completion of
the current interrupt.

Thus far, software interrupts were triggered by directly invoking the
kernel interrupt handler, i.e. they are executed in place within the
context of the current interrupt, for instance including locked mutexes.

Software interrupts are for instance used by the interrupt task
implementation to trigger the first task iteration. Since the functions
to issue an interrupt task and to retrieve the associated data have to
use the same MTX_NOINT mutex, the issue function has to release the
mutex before calling the interrupt handler, which otherwise would cause
a deadlock once the interrupt handler retrieves the task data. Releasing
the mutex is not an issue, however it requires the ksignal interface to
provide a wait function without an associated mutex.

As another side effect of calling the interrupt handler in place, the
ksignal associated to the interrupt task might be sent before calling
the wait function, which is thus far handled through buffering those
interim occurrences.

issue: 234