[ksignal: redefine API]
With this change the ksignal API is redefined, aiming to make the
ksignal-using code thread-safe more easily and reduce the code size and
complexity of the ksignal implementation.

Thus far ksignal_wait() and ksignal_send() where internally protected by
mutexes, not requiring their callers to protect them, causing the
following drawbacks:
- Signals could be sent before the associated waiting call was issued,
  potentially leading to a locked system. To prevent this, the ksignal
  implementation had to tracked those interim signals.
- ksignal_wait() offers multiple versions, with and without mutex
  arguments. Feature derivatives of ksignal_wait(), e.g. adding timeout
  functionality, would always need to also provide variants with and
  without mutexes.

With this change ksignal_wait() and ksignal_send() require the caller
to protect the calls with mutexes, leading to the following pattern:

	/* waiting thread */
	mutex_lock()

	// do stuff
	ksignal_wait()

	mutex_unlock()

	/* notifying thread */
	mutex_lock()

	// do stuff
	ksignal_send()

	mutex_unlock()

This should make it easier to get the calling code thread-safe. As a
first example the ioctl() implementation of netfs has been fixed, which
thus far did not unlock the related node's mutex before calling
ksignal_wait().

issue: 234