[itestfw: fix stuck system in blocking syscall]
issue:

System gets locked when issuing a blocking read syscall to a uart
which has interrupts disabled.

reason:

The kernel calls lnx_pause(), expecting an interrupt, e.g. the timer,
to wake it. However, only a couple of timer interrupts are issued before
the system gets stuck entirely.
After the first couple of timer interrupts the itestfw tries to issue a
scheduler interrupt, blocking the corresponding timer thread until the
kernel handled the scheduler interrupt. Since scheduler interrupt and
syscalls need to have the same priority and a syscall is already active,
the scheduler interrupt never gets issued to the kernel. Since both
interrupts, timer and scheduler, are triggered by the same, blocked
itestfw thread timer interrupts are not issued anymore either.

fix:

Since scheduler and syscall interrupts need to have the same priority to
avoid messing with the active brickos thread, it is not an option to
separate their priorities. Instead an additional itestfw thread is
introduced, such that timer and scheduler interrupts are issued by
separate threads.