[driver: fix tx interrupt completion]
issue:

Write operations wrongly report successful completion.

reason:

tx/write interrupt implementations don't wait for the final interrupt to
arise but report completion upon initiating the transmission.

This is not an issue if the underlying driver/hardware either cannot
cause tx errors or doesn't report them. However, in any other case a
transmission will be reported as a success even if an error occurs.

fix:

Extend the kernel interrupt tasks to check for the completion of a task
when querying its data. The driver can provide a callback to mark a task
as complete. If the callback indicates completion the interrupt task
implementation completes the task and queries for the next one.