[defconfig: hostbridge configurations]
Add default configurations for x86 and atmega1284p to build a
hostbridge, i.e. a bridge between and x86 and a target brickos instance
through a uart interface.
This also requires updates to the device tree. Since those changes
conflict with the existing device trees, new ones are added.

issue: 226