[i2c bridge: initials]
Add the i2c bridge interface and device drivers.

Besides adding the i2c bridge driver all configurations are updated and
new configurations for an i2c bridge for x86 and avr are added. The
device trees are updated accordingly.

The Kconfig for the bridge driver is updated to allow separate selection
of the device and interface parts of the bridge drivers to allow a
device tree to contain multiple bridges with the same id.