[bridge driver: fix errno handling]
issue:

If errno was used as the pointer that a bridge_read() would use, the
resulting errno value might not be the one written with bridge_write().

reason:

The bridge datagram rx() functions resets errno if it didn't already
return with an error. This behaviour was chosen to allow E_AGAIN during
bridge operations but not to communicate it to the caller.

fix:

Only reset errno in in rx() if E_AGAIN actually occurred.