[terminal: escape code support]
Add escape code interpretation to the terminal abstraction, controlled
through the terminal's CANON flag.

Further, the terminal driver interface is updated. Thus far the terminal
flags needed to be part of each terminal type, e.g. for uart or spi,
while the driver had to provide the offset into that type so that the
terminal functions could access the flags. With the update the low-level
driver provides its hardware-specific configuration to the terminal
abstraction, while configurations common to all terminals, such as the
flags, are provided by the terminal abstraction. The terminal
abstraction  also takes care to merge the hardware-specific and the
common configurations when requested by the user.