[test: fix test macro side-effect]
issue:

An expression with side-effect passed to TEST_INT_EQ() was evaluated
multiple times.

reason:

The side-effect expression was passed through the test macro's ref
parameter, which is not stored in a temporary variable but used multiple
times in the macro. Hence, the expression is evaluated more than once.

fix:

Introduce a temporary for the ref parameter.