[x86: kernel heap location]
Change the location of the kernel heap for the x86 architecture.

Thus far the kernel heap location was controlled through the linker
script and mapped to a static variable. In preparation for mmap and the
required overlay for x86, the kernel heap is now dynamically allocated
through a shared memory block.