[x86: mmap overlay]
Implement the x86 mmap overlay, based on a shared memory block for the
kernel heap between kernel and init.

To exchange the id of the shared memory block the "setup" x86 hardware
operation is added. The mmap overlay function in the kernel replaces the
actual address returned by the mmap syscall with the offset relative
to the kernel heap. The overlay function in the library applies this
offset to the base address of its kernel heap shared memory block and
returns the resulting address to the user.