[analyse: x86 boot lock]
It seems like a race condition which causes the itestfw to handle a
hardware-op from the kernel, which the kernel doesn't respond to, i.e.
the itestfw is stuck waiting for a response of the kernel.

It is unclear if the itestfw is handling an hardware-op that the kernel
never issued. A mixup with hardware-ops from init can be excluded since
the same behaviour can be observed if init is put to sleep for 15
seconds.