[x86: interrupt state cache]
Replace the x86 hardware-op HWO_INT_STATE with a interrupt state cache
within the kernel, i.e. the kernel tracks the interrupt state itself and
only issues interrupt state transition requests to the itestfw if the
desired state differs from the current.
With that HWO_INT_STATE is no longer required and is thus removed.