[x86: fix itestfw segfault]
issue:

The itestfw SEGFAULTs during kernel startup.

reason:

Race condition between kernel and the display thread initialisation.

The itestfw starts the child processes for kernel and init before
creating its own threads, such as the one responsible for the x11
display. Due to that the kernel might already issue a configure call to
the display, while the corresponding thread hasn't completed the
initialisation. As an effect the configure call uses uninitialised
variables, causing the SEGFAULT.

fix:

Delay configure calls while the initialisation is not complete.