[gpio driver: api change]
Add an intermediate interface in the gpio driver abstraction in order to
allow drivers to use the gpio interface without creating a devfs device.

Thus far gpio_create() created the gpio interface structure as well as a
devfs device.
To avoid creating a devfs device when only the interface is needed the
gpio abstraction layer is changed such, that gpio_create() only creates
the gpio structure and lower-level gpio drivers return it. To compensate
for the missing devfs device the gpio driver abstraction now provides
probe functions to register those.