[test coverage: fix coverage for unit tests]
Re-enable code coverage for unit tests on the host.

With 066f734c9a75b4a7feb1fe6085abf465fcb263ab code coverage was disabled
for host binaries, hence also for the unit test binary. Consequently,
code coverage for unit tests can only be acquired if the target binary
is executed.

With this change unit tests are run through the itestfw on the x86
architecture, i.e. using the target unit test binary. Since the itestfw
is not available for other architectures, the host binary is continued
to be used.