[unit tests: fix strerror() tests]
Fix the strerror() tests to account for the extended error reporting.

With 69f07e3fed91b99c11495a22299a4e58604c29c2 strerror() has been
extended to allow reporting the location of an error. However, the
respective unit tests have not been adapted, which was not detected
since the extended error reporting is disabled on the host.