[kernel: non-const sched_running()]
Update the return type of sched_running() and sched_running_nopanic(),
removing const.

Using const on the return type isn't meaningful since the sched_running
functions are the only means to get the current thread and in a lot of
places the purpose of acquiring it, is to modify it.