[x86: usignal overhaul]
Replace the x86 overlay-based usignal implementation.

Thus far usignals from the kernel were not forwarded to the application.
Instead the syscall overlays within the application were used. However,
with this only signals issued by the application could be handled while
those from the kernel were not.

With this change the overlay for sigsend is replaced with a new hardware
operation that forwards usignals from the kernel to the application.