[avr: fix interrupt routine]
issue:

Wrong numbers being printed with printf in the pinsampler and pwmctrl
init commands. Even though the variables contained correct values, wrong
ones are output.

reason:

The utoa_inv() function, used within the printf family, to convert
numbers to strings required division and modulo operations. Since there
is no hardware support for them, gcc provides them through libgcc.

Those libgcc functions use the r1 register, which according to the ABI,
is used as a zero register. For the interrupt routine, r1 was falsely
assumed to always be zero and thus using it as throw away register at
the isr exit, clearing it to zero afterwards. Consequently, if an
interrupt occurs while r1 is not zero the value is altered upon return
from the interrupt.

Interrupt routines needed r1 to pop the interrupt vector return address
from the stack. This address is pushed by the call instruction, which
constitutes the interrupt vector. The addresses is needed since all
vectors, except reset, use the same handler function. Hence, this
function needs to identify the number of the interrupt, which is
computed based on this return address.
However, this address is never actually used to return to and hence
needs to be popped from the stack before returning from the interrupt.
For this pop operation r1 is used as the target register.

fix:

Introduce a second level of interrupt handlers. The first level are the
hardware interrupt vectors which simply defer to the second level. This
is needed since the interrupt vectors have a fixed size, i.e. cannot
contain too many instructions.
The second level contains the call of the c interrupt handler. In
addition to the original call instruction, to push the interrupt vector
return address, it contains a return instruction, which pops the address
from the stack upon leaving the interrupt.

issue: 244