[buildsys: fix devictree rebuild]
issue:

Updating the kernel configuration does not rebuild the devtree on the
next call of make. Instead a second call is needed.

For instance, enabling a driver only rebuilds the kernel image on the
first call of make. While on second call the devtree and also the
kernel image are rebuild.

reason:

The devtree targets are missing a dependency to the configheader. Hence,
the header is build independently and also independently updates the
artificial fixdep headers. Consequently, updating one of the fixdep
headers does not impact the devtree targets in the same call of make.
The updated header is only detected on the next make call.

fix:

Add a dependency to the configheader for devtree targets.