[kernel: merge kernel and user heap]
Remove the separate heaps for kernel- and user-space and instead use one
pool of memory.

The split between kernel- and user-space heap is somewhat artificial and
not needed for any feature. A unified heap removes the need to guess the
demands of kernel and applications upfront, hence allows for a more
efficient use of the available memory.

On the downside, without any sorts of memory protection and a unified
heap it is more likely for miss-behaving applications to corrupt kernel
memory.