[x86: fix warning on RWX LOAD segment]
issue:

With gcc 12 a new warning appears when building the kernel for the x86
targets. The warning indicates that the kernel elf file contains a LOAD
segment which has rwx permissions set. Having such a segment is a
security vulnerability, since this enabled to modify the section data
and execute it.

reason:

The kernel linker script maps code and data to the same region. Hence,
the linker puts them into the same segment using rwx permissions, since
code needs to be executable and data needs to be read and writable.

fix:

Update the kernel linker script to map modifiable data to a newly
created kernel_data region. Hence, separating code and modifiable data.