[init: fix empty linker arrays]
issue:

init segfaults during application initialisation on the rp2040.

reason:

The memory between __data_start and __data_end does not contain all of
the .data section. Hence, some variables within the .data section are
not initialised during application startup. In particular the function
pointers for lib_init() calls are not initialised, causing the segfault
once those init calls are executed.

__data_start and __data_end do not enclose all of the .data section,
since two .data sections are present in the init binary and __data_start
and __data_end only enclose one of them. Hence, the content of the other
.data section never gets initialised.

Two .data sections are create due to the way how the linker arrays for
the init commands and user tests are implemented.
Depending on the brickos configuration, those arrays might be empty, in
which case the linker does not create their respective __start and
__stop pointers, which then causes the link to fail due to undefined
references. To avoid this the lists.lds linker script is added to the
link stage of init, ensuring the __start and __stop pointers always
exist.
However, the lists.lds approach only seemed to work. For the avr
architecture, only one .data section is created (probably due to a much
older gcc version), hence the __start and __stop pointers enclose the
entire .data section. For x86 multiple .data sections are created, but
since the architecture does not require any init code to prepare the
.data section content, this did not cause any problem. Only for the arm
architecture, for which multiple .data sections are generated and the
.data section content needs to be copied from flash to ram, the issue
showed.

fix:

It seems impossible to get the linker to merge multiple .data sections
if they originate from different linker scripts. Hence, instead of using
multiple linker scripts, provide one for applications, that allows to
define a plugin linker script, which can contain the additional .data
section content.

Since the application linker script is also intended to be used for
other, non-init, application binaries, it is not possible to include the
plugin linker script unconditionally, since with that all applications
would need to define it, regardless if they actually use it.
To also avoid maintaining multiple, largely identical, application
linker scripts, instead run the linker scripts through the compiler
preprocessor.