[buildsys: define char implementation]
Ensure char has the same range, representation and behaviour as signed
char.

The C standard leaves it up to the implementation to define if char-type
variables have a range, representation and behaviour according to signed
or unsigned char.

Thus far, the toolchains for avr and x86 treated char as signed char.
However, the arm toolchain treats it as unsigned char, which doesn't fit
the expectations for the code, raising some warnings.
Using -fsigned-char fixes gcc to always use signed char for char-type
variables.