[kernel: merge core and platform init]
Replace the core_init calls with more configurable platform_init calls.
In particular, allow to control whether a call shall be run on all cores
or only once on a single core.
Also remove leftover driver variables and comments.

Thus far dedicated core initialisation, that had to be run on all cores
before initialising anything else in the system, was sufficient. However,
with the arm v6m architecture, parts of the system/platform need to be
initialised in order to be able to complete the  core initialisation.
Specifically, the clock system has to be initialised, such that the clock
values in the devtree are updated and can be used to initialise the
systick timer, which there is a dedicated one per core.
If the clock values shall still be updated during initialisation through
measurement, rather than hard-coding them, a more flexible initialisation
is needed. Hence, making the order of core and platform initialisation
architecture-dependent.

To further avoid duplicating the check for PIR within each init call,
the singular vs. smp setting is added, moving the check from the
individual init calls to the kernel initialisation functions.