[syscall: transition kernel interface to sc_arg()]
Update the kernel internal syscall handling, moving the interrupt to the
kernel, keeping only a callback to retrieve the syscall arguments from
the architecture layer.

Previously, each architecture had to register a callback for the syscall
interrupt, implement the architecture-dependent retrieval of the syscall
arguments from user-space and call the kernel syscall handler.
To avoid duplicating code such as copy_to/from_user when accessing the
user-space arguments, the kernel now registers the syscall interrupt
handler, requiring the architecture layer to only provide a callback to
retrieve the user-space syscall pointer.

Further, add the syscall interrupt number to the architecture node of
the devtree to avoid requiring architectures to use the same name for
it in the architecture header.