[kernel statistics: discard]
Discard kernel-stat calls if kernel statistics are disabled.

If kernel-stat calls and regular kernel functions are part of the same
source file, the stat calls will always be part of the binary if the
object file is linked, even if kernel statistics are disabled.
To avoid this, a special .discard section is introduced and kernel-stat
calls are mapped to it, if kernel statistics are disabled, removing them
during linking.