[dtc: use ncores over core-mask]
Update the devtree script syntax to use the number of cores (ncores)
rather than core-mask.

Using the core-mask would for instance allow to specify to only use core
1 and disable core 0. For one there is currently no use-case to be that
specific. Secondly, not all architectures might allow this level of
control, e.g. core 0 might always be the boot core, which would belie
expectations, if one would rightly expect core 0 to never be running.
Further, all checks for (PIR == 0) would need to be updated, to rather
check for the first enabled core in the core-mask.