[mutex: fix corrupted interrupt state]
issue:

Core reporting disabled interrupts when calling sched_yield().

reason:

If, in a multi-core system, the contention for a NOINT mutex is high
enough, the interrupt state, saved by one core, can get picked up by
another one due to the order in which the interrupt state is saved and
the mutex lock is acquired.

    core 0                       core 1
mutex_lock
  state = int_disable
  acquire lock
       .                     mutex_lock
       .                       state = int_disable
       .                            .
mutex_unlock                        .
  release lock                      .
  int_reset(state)
                               acquire lock

The saved interrupt state is part of the mutex, hence shared among all
cores. In the above scenario this causes core 0 to restore the interrupt
state from core 1.

In case of the sched mutex, such a scenario causes a problem if the
following occurs:

    core 0                       core 1
kernel task
  call sched cleanup task
  acquire sched mutex
       .                     timer interrupt
       .                       disabled interrupts
       .                       try to acquire sched mutex
  release sched mutex
  call sched_yield

As described in the scenario above, core 0 enters the scheduler cleanup
task with enabled interrupts and picks up the disabled interrupt state
from core 1 when leaving it. Causing a kernel panic within sched_yield()
due to disabled interrupts.

fix:

mutex_lock:
  Ensure that the interrupt state is only updated within the mutex once
  the mutex is actually acquired.

mutex_unlock:
  Ensure that the interrupt state is restored based on the mutex state
  from before releasing the mutex.