[linker scripts: simplify alignment]
Update the linker scripts to avoid unaligned sections, e.g. when
adding support for a new architecture.

While adding support for arm-v6m and the rp2040 the kernel triggered a
hard fault when calling the first platform init function if the tests
for iskheap() are linked to the kernel image.
The hard fault was caused by an unaligned data access when loading the
address of the init function from the linker array. The address was not
aligned, since the strings that define a test, name and description,
changed the length of the regular .data section, causing the sections
for the kernel init calls to start at an unaligned address.

To fix this, the sections following the regular .data sections need to
be manually aligned in the linker script. This is the case for the x86
architecture and not needed for avr, since it is byte aligned. However,
for arm-v6m the alignment was missing.

To avoid missing the alignment in the future, the linker scripts are
extended, such that the included linker scripts ensure the needed
alignment using the information from the devtree. In order to allow
using the devtree header, all linker scripts are fed through the c
compiler pre-processor.