[kernel heap: fix iskheap()]
issue:

iskheap() classifies the kernel stack as being kernel heap, causing the
respective test to fail on avr and arm.

reason:

The kernel stack is cut from the end of the kernel heap devtree node.
However, iskheap() does not subtract the stack size when determining if
an address is part of the heap. Hence, if an address on a kernel stack
is passed to iskheap(), it is classified as heap.

On x86 the iskheap() test does not fail, since the kernel is separate
process, not using the kernel heap devtree node.

fix:

Update iskheap() to account for the kernel stack.

Also update the kernel heap initialisation to use the devtree macros
rather than dynamically looking up the heap devtree node. However, since
the x86 architecture dynamically allocates the kernel heap and thus has
to update the devtree node, the initial base address of the heap devtree
node is invalid and with it the value of the DEVTREE_HEAP_BASE macro.
To fix this, the devtree export is updated to make DEVTREE_HEAP_BASE
refer to the devtree node rather than the immediate value when building
for the x86 architecture.