[buildsys: fix undefined reference to memset]
issue:

When compiling init with -flto linking fails, reporting undefined
references to memset.

reason:

Seemingly, memset is removed through link time optimisation, since an
implementation is provided and the respective object files are linked.

Since memset is undefined even if it is used, it might be inlined by the
compiler and hence be removed as a global symbol. The locations causing
the undefined reference error are most likely the ones where the
compiler itself introduces memset.

fix:

Mark memset, memcpy and memcmp as used to prevent their symbols from
being removed.