[stream: fix negative number output for size_t]
issue:

Failing tests when printing negative numbers for length size_t on the
rp2040. Instead of, e.g. -10, a large integer, in this case 4294967286.
is printed.

reason:

When reading the variable to print size_t (unsigned) instead of ssize_t
(signed) is used as the type when reading the variable via va_arg().

The stream implementation uses only one variable to store integers of
different sizes. In case of the rp2040 size_t is smaller than the used
type for integer variables. Since the negative variable is loaded as
unsigned integer, which could be represented in the bigger integer type,
a large positive, instead of the intended negative number is printed.

fix:

Use the ssize_t to load the value via va_arg().