[vimgdb: overhaul startup script]
Overhaul the vimgdb startup script to make it easier to add support for
new architectures and targets that might have to the same architecture
but need a different startup routine for different devices.

With the change it is possible to add new targets based on the
CONFIG_ARCH and CONFIG_PROCESSOR variables.

Further, the loading of user scripts is simplified, no longer requiring
the individual dummy scripts, e.g. for breakpoints, while also allowing
the use a common user.vim script.