[Kconfig: move gpio integer size selection]
Move the selection of the gpio integer size from the user to the
platform driver.

Thus far the user could select the size of gpio integers. However, since
the value is hardware dependent, user interaction is not needed. Instead
now the underlying device driver has to specify the integer size.