[types: use _Bool]
Change the typedef for bool from uint8_t to the c standard's _Bool.

Using _Bool rather than uint8_t allows for bool-casts to actually cast a
value to 0 or 1. With uint8_t as the base time, 8-bit values would not
be changed when cast to bool. This is not an issue for logic operations
however, if bool casts are used for bit operations, e.g. to create a bit
mask, un-changed values would yield a wrong result.