[gpio configs: remove mode and pin_mask]
Remove the mode and pin_mask members from gpio_cfg_t.

Both members are used for controlling the sanity checks of a gpio
configuration. To simplify the interface both members are removed and
the sanity checks are implemented as devtree asserts. This reduces the
memory footprint of the devtree while still maintaining valid gpio
configurations.