[tests: run host unit tests on x86]
Also run the host unit test binary for the x86 architecture.

Thus far the host unit test binary was excluded for x86 since the target
unit test binary was executed. However, not all unit tests are included
in the target binary. Hence, it is beneficial to also run the host
binary.