[kernel: allow driver probe to fail]
Update the kernel initialisation to allow driver probe calls to fail
without a kernel panic.

Thus far a failing probe call resulted in aborting kernel initialisation
with a panic. However, since most drivers are not essential for the
kernel and system to work, a panic is not needed. Having an operational
system might also help analysing the system.