[kernel: fix interrupt fortell handler]
issue:

Not all foretold interrupts are handled before leaving the kernel
interrupt handler.

reason:

If a foretold interrupt has been handled the loop counters to iterate
the foretell data structure are reset to 0. However, since this is done
within a for loop body, the counter of the inner loop is incremented
immediately, causing the first part of the data structure being skipped,
missing the respective foretold interrupts.

fix:

Account for the counter increment, resetting the inner loop to -1.