[kernel: rename timer functions]
Rename the ktimer register and release functions to start and abort.

With a draft change, the ktimer_release() check if a timer already
expired had been removed, since it has been found that this check will
never be true, since all timers with a ticks value of zero would already
be removed by the timer handler.
While this is true, ktimer_release() can, and in fact is, called after
a timer expired, cf. ksignal_timedwait(). In that case the ticks value
will be zero and the timer must not be removed a second time.

By renaming the functions and adding a respective comment the intention
is hopefully clearer and removing the condition again can be avoided.