[threads: fix locking]
issue:

Potential system lockup.

reason:

thread_ctx_pop(), which is responsible to deliver signal to threads, is
called on interrupt exit. The thread signal list is protected by a
normal mutex. That is, if an interrupt occurs while the mutex is locked,
the following thread_ctx_pop() would lockup the system.
This could for instance happen if an interrupt occurs while adding a new
signal to the list.

fix:

Disable interrupts while having the thread mutex locked by using the
MTX_NOINT mutex attribute.