[arm v6m: fix exception priorities]
issue:

On the rp2040 the kernel reports an uncaught error during syscall
execution, i.e. errno indicating an error while the syscall's return
value doesn't.

reason:

Actually the syscall didn't encounter an error, but the reported errno
value is the default (E_UNKNOWN) used when triggering a syscall in order
to detect if the kernel updates the syscall parameter struct. E_UNKNOWN
is currently only used for that purpose, hence the kernel did not update
the data structure. In fact, the kernel updated the syscall parameters
of another thread twice instead.

The wrong context being used is caused by the syscall implementation on
the arm v6m. User-space syscalls use the svcall instruction, while the
kernel has to use the pendsv exception, which is configure to have the
higher priority then the svcall exception. This allows the kernel to
trigger syscalls from within user-space syscall handlers, which is for
instance required to suspend a thread through sched_yield().
In contrast to svcall, pendsv is interrupt-driven, i.e. the program sets
the ICSR[PENDSVSET] bit, making the pendsv exception pending, then
waiting for the exception being handled. However, the exception is not
necessarily taken immediately. That is, pendsv is asynchronous to the
program execution and it is possible for another exception, with higher
priority, to be handled after making pendsv pending but before taking
the exception handler.

In particular the following scenario caused the wrong thread context to
be used:
 - user-thread triggers a sleep syscall
 - kernel suspends the thread and returns to kernel-thread

   ...

 - user-thread is made ready again by a systick (but not scheduled yet)

   ...

 - kernel calls sched-yield syscall, i.e. triggering a pendsv exception
 - after making pendsv pending but before taking the exception a systick
   exception occurs
 - since systick has higher priority it is handled before pendsv
 - during the systick the user-thread is scheduled, i.e. is now the
   active thread
 - upon returning from the systick, instead of returning to the user-
   thread, the pendsv handler is called, since its exception is still
   pending
 - the related syscall handler retrieves the syscall parameter from
   the active thread, which now is the user-thread and not the kernel-
   thread that triggered the syscall
 - since both threads requested the same syscall (sched-yield), the
   correct syscall is triggered but the return value is written to
   user-space
 - upon completion of the pendsv handler, execution returns to the
   user-thread, i.e. in the sched-yield handler, updating the syscall
   parameters a second time

   ...

 - when returning to the kernel-thread, i.e. to the point after
   triggering the pendsv exception, it returns the unchanged errnum
   attribute and the kernel syscall handler reports the uncaught syscall
   error

This cannot happen with svcall, since it is synchronous to program
execution. That is, a systick might either occur before svcall is
executed or after it has been executed, i.e. the exception is already
processed. The first case is not problematic, since the syscall has not
even been started. The second case is also not a problem since the
systick would return to the user-thread rather than the svcall exception
handler, since the exception is already being processed, i.e. not in the
intermediate state between getting pending and entering the exception
handler.

fix:

Raise the pendsv exception priority over the one from systick, ensuring
pendsv exceptions are handled before systick exceptions.