[rp2040: add default pad configurations]
Update the rp2040 platform initialisation, adding support for default
pad configurations based on the selected pad function.

Some peripherals, such as i2c, do not configure the i/o pads themselves
but have certain requirements towards the configuration.
This could be handled in the i2c driver. However, that would require to
provide the allocated pads to the driver, i.e. the allocation from pad
to function would be spread over multiple parts of the devtree, making
it harder to get a concise view of the final pad configuration.
To avoid this, any function-specific pad configurations are made by the
platform initialisation.