[i2c: fix avr baudrate calculation]
issue:

When trying to connect an atmega1284p master to a raspberry pi pico
slave the i2c driver returns with E_NOCONN even though the target slave
is connected and addressable.

reason:

The baudrate calculation in the avr i2c driver is faulty. In particular
the multiplication of the target value (100kHz) by 1000 to convert it to
Hz, overflows since the target value is a 16 bit integer. Hence, the
resulting value doesn't reflect the desired target clock.

fix:

Reorder the operations to prevent the overflow.