[i2c: make i2c mode reconfigurable]
Update the i2c driver api, moving the decision whether to use master or
slave mode to the higher-level device driver instead of hard-coding it
in the devtree.

Main purpose for this change is to ease interactive testing and
debugging of i2c drivers since the change allows to reconfigure the i2c
device without reflashing the device.