[i2c: merge master and slave command lists]
Merge the i2c master and slave command lists to simplify the driver
implementation.

The differentiation was somewhat artificial since previously a device
was hard-coded to either master or slave mode via the devtree.

With the recent changes both modes are possible, however for now it is
sufficient to assert an error if hardware state and datagram do not
match, since the slave implementation will be changed to a
non-user-triggered implementation anyways.