[i2c: reflect number of bytes transfered]
Update the i2c driver api, in particular the return values of the i2c
read/write/xfer functions, to pass the number of bytes transfered
through to the caller.

This is not used by most of the higher-level drivers, except the raw i2c
device driver, which needs it to correctly communicate to user-space.