[i2c: add acked driver callback]
Extend the i2c driver interface, adding the acked callback.

The callback is added for i2c hardware that is able to issue/stage more
write operations than what might actually be finished, i.e. accepted by
the remote master or slave. Currently, this is a preparation for the
rp2040.