[i2c: greedy datagram handler]
Update the i2c datagram handler to not return while the underlying i2c
hardware driver is reporting new states.

This behaviour is needed if the underlying i2c hardware does not require
software interaction for all i2c states which are used by the i2c
protocol handler. For instance, if the hardware automatically issues a
start condition when connecting to a slave, the lower-level driver still
has to report the start state, in order to allow the i2c protocol handler
to correctly update its internal state. However, if the hardware does not
require software interaction for all state transitions, it also will not
trigger interrupts, which would usually invoke the next interaction of
the protocol handler, in those cases. Hence, the protocol implementation
has to poll for software-only state transitions.

An alternative solution would be to foretell interrupts. However, this
would add
 - more complexity to the lower-level drivers
 - probably duplication between such drivers
 - a higher delay between protocol state transitions, since the
   interrupt handler would first return to the kernel interrupt handler,
   which would then handle the foretold interrupt.