[i2c: add interface device]
Introduce the i2c bus device which handles the i2c protocol layer
implementation, including the hardware configuration and control as well
as the creation of the interface used by the higher-level i2c device
drivers.

This is an attempt to unify the device driver abstraction levels, having
a low-level hardware driver, a kernel level interface driver and higher
level drivers.